rename: fixture 패키지를 util 폴더 내부로 이동
