+++ be/promotion/src/test/java/woowa/promotion/acceptance/admin/AdminAcceptanceTest.java
@@ -8,8 +8,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import woowa.promotion.fixture.FixtureFactory;
 import woowa.promotion.util.AcceptanceTest;
+import woowa.promotion.util.fixture.FixtureFactory;
 
 @DisplayName("[인수테스트][관리자] 인증")
 public class AdminAcceptanceTest extends AcceptanceTest {
