+++ be/promotion/src/test/java/woowa/promotion/application/admin/promotion/application/PromotionServiceTest.java
@@ -1,17 +1,17 @@
 package woowa.promotion.application.admin.promotion.application;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static woowa.promotion.fixture.CouponGroupFixture.A_쿠폰그룹;
-import static woowa.promotion.fixture.CouponGroupFixture.B_쿠폰그룹;
-import static woowa.promotion.fixture.FixtureFactory.createCouponGroup;
-import static woowa.promotion.fixture.FixtureFactory.createPromotion;
-import static woowa.promotion.fixture.FixtureFactory.createPromotionOption;
-import static woowa.promotion.fixture.FixtureFactory.createPromotionOptionRequest;
-import static woowa.promotion.fixture.FixtureFactory.createPromotionRegisterRequest;
-import static woowa.promotion.fixture.PromotionFixture.A_프로모션;
-import static woowa.promotion.fixture.PromotionFixture.B_프로모션;
-import static woowa.promotion.fixture.PromotionOptionFixture.A_프로모션_옵션;
-import static woowa.promotion.fixture.PromotionOptionFixture.B_프로모션_옵션;
+import static woowa.promotion.util.fixture.CouponGroupFixture.A_쿠폰그룹;
+import static woowa.promotion.util.fixture.CouponGroupFixture.B_쿠폰그룹;
+import static woowa.promotion.util.fixture.FixtureFactory.createCouponGroup;
+import static woowa.promotion.util.fixture.FixtureFactory.createPromotion;
+import static woowa.promotion.util.fixture.FixtureFactory.createPromotionOption;
+import static woowa.promotion.util.fixture.FixtureFactory.createPromotionOptionRequest;
+import static woowa.promotion.util.fixture.FixtureFactory.createPromotionRegisterRequest;
+import static woowa.promotion.util.fixture.PromotionFixture.A_프로모션;
+import static woowa.promotion.util.fixture.PromotionFixture.B_프로모션;
+import static woowa.promotion.util.fixture.PromotionOptionFixture.A_프로모션_옵션;
+import static woowa.promotion.util.fixture.PromotionOptionFixture.B_프로모션_옵션;
 
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
