+++ be/promotion/src/test/java/woowa/promotion/application/app/promotion/application/AppPromotionServiceTest.java
@@ -7,9 +7,9 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import woowa.promotion.app.promotion.application.AppPromotionService;
-import woowa.promotion.fixture.FixtureFactory;
-import woowa.promotion.fixture.PromotionFixture;
 import woowa.promotion.util.ApplicationTest;
+import woowa.promotion.util.fixture.FixtureFactory;
+import woowa.promotion.util.fixture.PromotionFixture;
 
 @DisplayName("[비즈니스 로직 테스트] 회원 프로모션")
 class AppPromotionServiceTest extends ApplicationTest {
