+++ be/promotion/src/test/java/woowa/promotion/util/fixture/CouponGroupFixture.java
@@ -1,4 +1,4 @@
-package woowa.promotion.fixture;
+package woowa.promotion.util.fixture;
 
 import java.time.Instant;
 import java.time.temporal.ChronoUnit;
